+++ benchmark/dense_sparse/decompression.cpp
@@ -43,7 +43,7 @@ void readHeader(size_t** cLengths, void* const cBuff, size_t cSize){
     chunkSize = (size_t)strtol(token, &ptr, 10);
 
     *cLengths = (size_t*)malloc_orDie(sizeof(size_t) * (numOfChunks + 1));
-    for(int i = 0; i < numOfChunks; i++){
+    for(size_t i = 0; i < numOfChunks; i++){
         token = strtok(NULL, delim);
         headerSize += strlen(token) + 1;
         (*cLengths)[i] = (size_t)strtol(token, &ptr, 10);
@@ -69,7 +69,7 @@ static void decompress(const char* fname, const char* oname, const ZSTD_DDict* d
     // cLengths stores the sizes of the compressed blocks
     size_t* cLengths;
     readHeader(&cLengths, cBuff, cSize);
-    int offset = headerSize;
+    size_t offset = headerSize;
     const size_t denseSize = numOfChunks * threshold;
     
     /* Check that the dictionary ID matches.
@@ -96,7 +96,7 @@ static void decompress(const char* fname, const char* oname, const ZSTD_DDict* d
     randomAccessTime = 0, totalTime = 0;
     clock_t begin = clock();
 
-    for(int chunk = 0; chunk < numOfChunks; chunk++){
+    for(size_t chunk = 0; chunk < numOfChunks; chunk++){
         // Size of the compressed block (obtained from the header)
         size_t cBlockSize = cLengths[chunk];
         // To store together the dense and sparse blocks of the current chunk
