+++ src/basic/dictionary_decompression_chunks.cpp
@@ -15,10 +15,10 @@
 #include <zstd.h>      // presumes zstd library is installed
 #include "../common.h"    // Helper functions, CHECK(), and CHECK_ZSTD()
 
-static int numOfChunks, headerSize = 0;
+static size_t numOfChunks, headerSize = 0;
 
 struct entry{
-    int cPos, inPos, inSize;
+    size_t cPos, inPos, inSize;
 };
  
 /* createDict() :
@@ -43,7 +43,7 @@ struct entry* readHeader(void* const cBuff, size_t cSize){
 
     struct entry* table = (struct entry*)malloc_orDie(sizeof(struct entry) * (numOfChunks + 1));
 
-    for(int cnt = 0; cnt < numOfChunks; cnt++){
+    for(size_t cnt = 0; cnt < numOfChunks; cnt++){
         token = strtok(NULL, delim);
         table[cnt].cPos = strtol(token, &ptr, 10);
         headerSize += strlen(token) + 1;
@@ -91,13 +91,13 @@ static void decompress(const char* fname, const char* oname, const ZSTD_DDict* d
     ZSTD_DCtx* const dctx = ZSTD_createDCtx();
     CHECK(dctx != NULL, "ZSTD_createDCtx() failed!");
 
-    int offset = headerSize;
+    size_t offset = headerSize;
     unsigned long long outSize = 0; 
-    table[numOfChunks] = (struct entry){cSize - offset, table[numOfChunks - 1].inPos + table[numOfChunks - 1].inSize, 0};
+    table[numOfChunks] = (struct entry){cSize - offset, table[numOfChunks - 1].inPos + table[numOfChunks - 1].inSize, 0ll};
 
     void* const out = malloc_orDie(4ll * (long long)cSize); // Assuming the max. compression ratio is 4
 
-    for(int chunk = 0, pos = offset + table[0].cPos; chunk < numOfChunks; chunk++){
+    for(size_t chunk = 0, pos = offset + table[0].cPos; chunk < numOfChunks; chunk++){
         size_t chunkSize = table[chunk + 1].cPos - table[chunk].cPos;
         unsigned long long const rSize = ZSTD_getFrameContentSize((unsigned char*)cBuff + pos, chunkSize);
         CHECK(rSize != ZSTD_CONTENTSIZE_ERROR, "%s: not compressed by zstd!", fname);
